Task_3/requirements.txt
torch
torchvision
numpy
pillow
scikit-learn
pandas
